move sorting logic to hook
